publish jdcloud-sdk-go version 0.0.10

+++ services/starshield/models/DateHistogram.go
@@ -27,7 +27,7 @@ type DateHistogram struct {
 如果是流量，数据点的单位是Byte
 如果是请求量，数据点的单位是次数
  (Optional) */
-    Timeseries []int `json:"timeseries"`
+    Timeseries []float64 `json:"timeseries"`
 
     /*  (Optional) */
     Unit string `json:"unit"`
@@ -36,5 +36,5 @@ type DateHistogram struct {
 如果是流量，它的单位是Byte，代表数据点流量之和。
 如果是请求量，它的单位是次数，代表数据点请求量之和。
  (Optional) */
-    Total int `json:"total"`
+    Total float64 `json:"total"`
 }
